Pass flags to fcntl(F_SETFL) as int, not long
Everybody but Linux documents this as taking an int, and Linux is
very tolerant of getting an int instead.  If it weren't, everybody
doing fcntl(fd,F_SETFL,O_NONBLOCK) would break, since the glibc
headers define O_NONBLOCK as an int literal.

+++ evutil.c
@@ -169,7 +169,7 @@ evutil_make_socket_nonblocking(int fd)
 	}
 #else
 	{
-		long flags;
+		int flags;
 		if ((flags = fcntl(fd, F_GETFL, NULL)) < 0) {
 			event_warn("fcntl(%d, F_GETFL)", fd);
 			return -1;
